vrank: move getter tests and add pf report coverage
